backend/requirements.txt
# Core dependencies
yt-dlp>=2024.1.0
python-dotenv>=1.0.0
mutagen>=1.47.0

# CLI framework
typer>=0.12.0
rich>=13.0.0

# Terminal utilities
termcolor>=2.3.0

# Optional: GCP upload
google-cloud-storage>=2.10.0

# Build tool
pyinstaller>=6.0.0

